helpers.xr_utils now keeps track of accessors it has registered.
Included an internal dict to register what classes have been registered as
xarray accessors and under which names, as well as functions to retrieve entries
from that dictionary.

I have still not added functionality to avoid re-registering accessors that have
already been registered, but this should be done (i.e., make the `register_`
functions check whether the class in question has already been registered under
the requested name, and if so avoid registering it again, or perhaps raise an
error if it has already been registered under a different name).