Added dataload files (just copied from climate_trace_comparison repo).
Added package files for loading and processing raw data files. These files are
currently just copied from climate_trace_comparison, and not consistent with
anything that is done in cmdata.datadesc or anywhere else in the cmdata package.
Just putting it here to be able to use and modify the code as needed in a
centrally managed package.