Added import error handling in import of cmplot in cmdata package.
Importing the main cmdata package also automatically imports cmdata.cmplot and
cmdata.cmplot.hv_plotting.  Previously, this would trigger an ImportError if any
of the plotting package dependencies were not installed. With the new behavior,
the error is caught and ignored. The submodules/subpackages that could not be
imported are then not present in the cmdata.cmplots package object, and instead
the import errors are stored in the dict cmdata.cmplot.submodule_import_errors
(with the names of the modules/packages as keys).